% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_map.R
\name{cor_map}
\alias{cor_map}
\title{Correlation analysis of a dataset}
\usage{
cor_map(
  dataset,
  select_cols = colnames(dataset),
  filter_rows = NULL,
  metadata = NULL,
  self = "yes",
  method = "BH"
)
}
\arguments{
\item{dataset}{A \code{\link[data.table:data.table]{data.table}}. Must be in
column-major order.}

\item{select_cols}{A vector of column names (character), or indices
(numeric). The columns to use for computing correlations, which must be of
type numeric.}

\item{filter_rows}{A named \code{\link[base:list]{list}}. Values specify
which rows to subset. Names correspond to column names in \code{dataset}, or
\code{metadata} if supplied.}

\item{metadata}{A \code{\link[data.table:data.table]{data.table}}. Must be
in column-major order. Optional input containing data to filter
\code{dataset} by. If supplied, \code{metadata} and \code{dataset} must
possess the same \code{\link[data.table:setkey]{key}} column.}

\item{self}{A character scalar. Self-correlations are included if
\code{"yes"} (default), or omitted if \code{"no"}.}

\item{method}{A character scalar. Correction method for p-value adjustment,
passed to \code{\link[stats:p.adjust]{p.adjust}}. One of: \code{c("holm",
"hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")}; \code{"BH"}
(alias \code{"fdr"}) (default).}
}
\value{
A \code{\link[data.table:data.table]{data.table}} in long-format of
of Pearson's product moment correlation coefficients (\code{r}), p-values
(\code{p}), adjusted p-values (\code{q}), and observation counts
(\code{n}) for pairwise Pearson's correlations.
}
\description{
Computes pairwise Pearson's correlations of a dataset using
\code{\link[collapse:pwcor]{collapse::pwcor}}, with the ability to filter
and subset the input. Returns a long-format
\code{\link[data.table:data.table]{data.table}} of Pearson's product moment
correlation coefficients, p-values, adjusted p-values, and observation
counts.
}
\section{Subset}{

\code{\link{cor_map}} conveniently allows filtering (\code{filter_rows}) of
the input \code{dataset} by performing a cross-join
(\code{\link[data.table:CJ]{CJ}}) with a named \code{\link[base:list]{list}}
referring to values present within the \code{dataset} itself, or a separate
\code{metadata} \code{\link[data.table:data.table]{data.table}}. If the
\code{dataset} contains non-numeric columns they must be omitted by
selecting (\code{select_cols}) the columns to compute pairwise correlations.
This mechanism also allows limiting of the correlations to perform. The
subsetting algorithm is identical in \code{\link{cor_target}} and
\code{\link{cor_target_map}}.
}

\section{Correlation analysis}{

\code{\link{cor_map}} uses \code{\link[collapse:pwcor]{collapse::pwcor}} to
test for an association between paired samples computed with Pearson's
product moment correlation coefficient (same as
\code{\link[stats:cor.test]{cor.test}}). Correlations are computed on
complete pairs of observations for each pair of variables (same as
\code{\link[stats:cor]{cor(..., use = "pairwise.complete.obs")}}).

Adjusted p-values are computed with \code{\link[stats:p.adjust]{p.adjust}}
using the one of the \code{\link[stats:p.adjust.methods]{p.adjust.methods}}:
\code{c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr",
"none")}. Default \code{"BH"} (alias \code{"fdr"}) is the Benjamini &
Hochberg (1995) false discovery rate multiple testing adjustment method.
}

\section{Output}{

A \code{\link[data.table:data.table]{data.table}} in long-format is returned
with no pairwise duplicates. If \code{filter_rows} has been supplied, the
filters are included.
}

\examples{
# Input
mtcars_DT <- data.table::as.data.table(mtcars, keep.rownames = "id")
data.table::setkey(mtcars_DT, id)

# Correlation analysis
cor_map(mtcars_DT, select_cols = colnames(mtcars_DT[, !("id")]))
}
\seealso{
\itemize{
\item \code{\link{cor_target}} for correlation analysis of a single target.
\item \code{\link{cor_target_map}} for correlation analysis of multiple targets.
}
}
