% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_target.R
\name{cor_target}
\alias{cor_target}
\title{Correlation analysis of a single target}
\usage{
cor_target(
  dataset,
  target,
  select_cols = colnames(dataset),
  filter_rows = NULL,
  metadata = NULL,
  self = "yes",
  method = "BH"
)
}
\arguments{
\item{method}{}
}
\value{

}
\description{
Computes pairwise Pearson's correlations of a single target using
\code{\link[stats:cor.test]{cor.test}}, with the ability to filter and
subset the input. Returns a long-format
\code{\link[data.table:data.table]{data.table}} of Pearson's product moment
correlation coefficients, p-values and adjusted p-values.
}
