% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_parallel.R
\name{end_parallel}
\alias{end_parallel}
\title{Create a sequential future}
\usage{
end_parallel()
}
\value{
A \code{\link[future:sequential]{SequentialFuture}}.
}
\description{
Begins a sequential \code{\link[future:plan]{future::plan}}, replacing
any previously set \code{\link[future]{plan}}.
}
\details{
\code{\link{end_parallel}} creates a sequential \code{\link[future]{plan}}
for synchronous processing in the current \code{R} session.

Run \code{\link{end_parallel}} at the end of working with \code{cordial}
functions to return to sequential processing. Alternatively, users may set
their own \code{\link[future]{plan}} explicitly for more control.
}
\examples{
# Begin sequential processing

end_parallel()
}
\seealso{
\itemize{
\item \code{\link{start_parallel}} for parallel processing.
\item \code{\link{cor_target}} for correlation analysis of a single target.
\item \code{\link{cor_target_map}} for correlation analysis of multiple targets.
\item \code{\link[future:plan]{future::plan}} for setting the future plan
directly.
}
}
