% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_parallel.R
\name{start_parallel}
\alias{start_parallel}
\title{Start a parallel multisession future}
\usage{
start_parallel(logical_cores = FALSE)
}
\arguments{
\item{logical_cores}{A logical scalar. The number of parallel \code{R}
sessions is set to the number of physical CPUs/cores if \code{FALSE}
(default), or logical CPUs/cores if \code{TRUE}.}
}
\value{
A \code{\link[future:multisession]{MultisessionFuture}}.
}
\description{
Begins a multisession \code{\link[future:plan]{future::plan}} for
parallel processing, replacing any previously set
\code{\link[future]{plan}}.
}
\details{
\code{\link{start_parallel}} creates a multisession
\code{\link[future]{plan}} for asynchronous (parallel) processing in
separate \code{R} sessions running in the background of the same machine.

If a multisession \code{\link[future]{plan}} is set,
\code{\link{cor_target}} and \code{\link{cor_target_map}} execute in
parallel; the result is returned to the main \code{R} session.

Run \code{\link{start_parallel}} at the beginning of working with
\code{cordial} functions. Alternatively, users may set their own
\code{\link[future]{plan}} for more control. To return to sequential
processing, see \code{\link{end_parallel}}.
}
\examples{
# Create parallel multisessions equal to the number of physical cores

start_parallel()       # Default
start_parallel(FALSE)  # Same as above


# Create parallel multisessions equal to the number of logical cores

start_parallel(TRUE)
}
\seealso{
\itemize{
\item \code{\link{end_parallel}} for sequential processing.
\item \code{\link{cor_target}} for correlation analysis of a single target.
\item \code{\link{cor_target_map}} for correlation analysis of multiple targets.
\item \code{\link[future:plan]{future::plan}} for setting the future plan
directly.
}
}
